+++ app/Http/Controllers/ServerController.php
@@ -40,6 +40,8 @@ class ServerController extends Controller
                 'type' => config('app.env'),
                 'search_host' => $searchHost,
                 'search_version' => $searchVersion,
+                'is_feeds_enabled' => (bool) config('bar-assistant.enable_feeds') === true,
+                'is_password_login_enabled' => config('bar-assistant.enable_password_login') === true,
             ]
         ]);
     }