+++ config/bar-assistant.php
@@ -69,5 +69,9 @@ return [
     'metrics' => [
         'enabled' => (bool) env('METRICS_ENABLED', false),
         'allowed_ips' => explode(',', env('METRICS_ALLOWED_IPS', '')),
-    ]
+    ],
+
+    'enable_feeds' => (bool) env('ENABLE_FEEDS', false),
+
+    'enable_password_login' => (bool) env('ENABLE_PASSWORD_LOGIN', true),
 ];